IOMAD: added filtering for course custom fields to reports/other pages. Moved forms for this to standard classes folder and out of user.php
