+++ local/iomad/lang/en/local_iomad.php
@@ -35,7 +35,9 @@ $string['companycourses'] = 'Other company courses';
 $string['companyfilter'] = 'Filter results';
 $string['companynamefilter'] = 'Company name contains';
 $string['companysearchfields'] = 'Company search fields';
+$string['coursesearchfields'] = 'Course search';
 $string['crontask'] = 'IOMAD Cron';
+$string['datesearchfields'] = 'Date search';
 $string['emailfilter'] = 'Email address contains';
 $string['firstnamefilter'] = 'First name contains';
 $string['iomad'] = 'IOMAD';
@@ -64,5 +66,5 @@ $string['removelicenses'] = 'Deleted - Company course records and licenses';
 $string['setupiomad'] = 'Start setting up IOMAD';
 $string['show_suspended_users'] = 'Show suspended users?';
 $string['userfilter'] = 'Filter results';
-$string['usersearchfields'] = 'User search fields';
+$string['usersearchfields'] = 'User search';
 $string['show_suspended_companies'] = 'Show suspended companies?';
