+++ local/report_completion_monthly/styles.css
@@ -1,15 +1,15 @@
-#page-local-report_completion_monthly-index .col-md-9 {
+#page-local-report_completion_monthly-index .col-md-9, #page-local-report_completion_monthly-index .chart-area {
     clear:both;
 }
 
-#page-local-report_completion_monthly-index #iomad_course_selector, #page-local-report_completion_monthly-index .form-group {
+#page-local-report_completion_monthly-index .coursessearchform, #page-local-report_completion_monthly-index .iomaddatesearchform {
     float:left;
 }
 
-#page-local-report_completion_monthly-index #iomad_course_selector select , #page-local-report_completion_monthly-index .felement {
+/*#page-local-report_completion_monthly-index #iomad_course_selector select , #page-local-report_completion_monthly-index .felement {
     background-color: #ddd;
 }
-
+*/
 #page-local-report_completion_monthly-index #iomad_course_selector, #page-local-report_completion_monthly-index .femptylabel {
     padding-top: 5px;
     padding-bottom: 5px;
@@ -18,4 +18,54 @@
 
 #page-local-report_completion_monthly-index .femptylabel .felement{
     background-color: #fff;
+}
+
+.iomadcoursesearchform .icons-collapse-expand.collapsed .expanded-icon,
+.iomadcoursesearchform .icons-collapse-expand.collapsed .collapsed-icon,
+.iomaddatesearchform  .icons-collapse-expand.collapsed .expanded-icon,
+.iomaddatesearchform .icons-collapse-expand.collapsed .collapsed-icon{
+    font-size: small;
+}
+
+.iomadcoursesearchform h3,
+.iomaddatesearchform h3 {
+    font-size: medium;
+}
+
+.iomadcoursesearchform .mform .form-inline .form-control,
+.iomadcoursesearchform .mform .form-inline .custom-select,
+.iomaddatesearchform .mform .form-inline .form-control,
+.iomaddatesearchform .mform .form-inline .custom-select {
+    max-width:75%;
+}
+
+.iomadcoursesearchform,
+.iomaddatesearchform,
+.iomadusersearchform {
+    margin-right: 10px;
+    float: left;
+}
+
+.iomadcoursesearchform .icons-collapse-expand.collapsed .expanded-icon,
+.iomadcoursesearchform .icons-collapse-expand.collapsed .collapsed-icon,
+.iomaddatesearchform  .icons-collapse-expand.collapsed .expanded-icon,
+.iomaddatesearchform .icons-collapse-expand.collapsed .collapsed-icon,
+.iomadusersearchform  .icons-collapse-expand.collapsed .expanded-icon,
+.iomadusersearchform .icons-collapse-expand.collapsed .collapsed-icon {
+    font-size: small;
+}
+
+.iomadcoursesearchform h3,
+.iomaddatesearchform h3,
+.iomadusersearchform h3 {
+    font-size: medium;
+}
+
+.iomadcoursesearchform .mform .form-inline .form-control,
+.iomadcoursesearchform .mform .form-inline .custom-select,
+.iomaddatesearchform .mform .form-inline .form-control,
+.iomaddatesearchform .mform .form-inline .custom-select,
+.iomadusersearchform .mform .form-inline .form-control,
+.iomadusersearchform .mform .form-inline .custom-select {
+    max-width:75%;
 }
