+++ local/report_completion_overview/index.php
@@ -39,6 +39,7 @@ $perpage      = optional_param('perpage', $CFG->iomad_max_list_users, PARAM_INT)
 // Id of user to tweak mnet ACL (requires $access).
 $acl          = optional_param('acl', '0', PARAM_INT);
 $search      = optional_param('search', '', PARAM_CLEAN);// Search string.
+$coursesearch = optional_param('coursesearch', '', PARAM_CLEAN);// Search string.
 $departmentid = optional_param('deptid', 0, PARAM_INTEGER);
 $courses = optional_param_array('courses', NULL, PARAM_INTEGER);
 $licenseid    = optional_param('licenseid', 0, PARAM_INTEGER);
@@ -83,6 +84,9 @@ if ($bycourse) {
 if ($search) {
     $params['search'] = $search;
 }
+if ($coursesearch) {
+    $params['coursesearch'] = $coursesearch;
+}
 if ($departmentid) {
     $params['deptid'] = $departmentid;
 }
@@ -112,6 +116,21 @@ if ($sort == "name") {
 }
 
 require_login();
+// Get course customfields.
+$usedfields = [];
+$customfields = $DB->get_records_sql("SELECT cff.* FROM
+                                      {customfield_field} cff 
+                                      JOIN {customfield_category} cfc ON (cff.categoryid = cfc.id)
+                                      WHERE cfc.area = 'course'
+                                      AND cfc.component = 'core_course'
+                                      ORDER BY cfc.sortorder, cff.sortorder");
+foreach ($customfields as $customfield) {
+    ${'customfield_' . $customfield->shortname} = optional_param('customfield_' . $customfield->shortname, null, PARAM_ALPHANUMEXT);
+    if (!empty(${'customfield_' . $customfield->shortname})) {
+        $params['customfield_' . $customfield->shortname] = ${'customfield_' . $customfield->shortname};
+        $usedfields[$customfield->id] = ${'customfield_' . $customfield->shortname};
+    }
+}
 
 $systemcontext = context_system::instance();
 
@@ -223,22 +242,59 @@ if ($departmentid == 0 ) {
     $departmentid = $userhierarchylevel;
 }
 
-$coursesform = new local_iomad\forms\course_select_form($linkurl, $params);
-
+$coursesform = new \local_iomad\forms\course_search_form($linkurl, $params);
 // Deal with company courses and search.
 $allcompanycourses = $company->get_menu_courses(true, false, false, false, false);
 $courselistsql = "";
+$coursesearchparams = [];
 if (!empty($allcompanycourses)) {
     $courselistsql = " AND ic.courseid IN (" . implode(',', array_keys($allcompanycourses)) . ")";
 }
 if ($showexpiryonly) {
-    $courselistsql .= " AND ic.validlength > 0";
+    $courselistsql = " AND ic.validlength > 0";
 }
+
+// Course name search.
+if (!empty($coursesearch)) {
+    $courselistsql .= " AND " . $DB->sql_like('c.fullname', ':coursename', false, false);
+    $coursesearchparams['coursename'] = "%" . $coursesearch . "%";
+}
+
+// Deal with any custom course field searches.
+$fieldcourseids = [];
+if (!empty($usedfields)) {
+    $foundfields = [];
+    foreach ($usedfields as $fieldid => $fieldsearchvalue) {
+        if ($customfields[$fieldid]->type == 'text' || $customfields[$fieldid]->type == 'text' ) {
+            $fieldsql = "fieldid = :fieldid AND " . $DB->sql_like('value', ':fieldsearchvalue');
+            $fieldsearchvalue = '%' . $fieldsearchvalue . '%';
+        } else {
+            $fieldsql = "value = :fieldsearchvalue AND fieldid = :fieldid";
+        }
+        $foundfields[] = $DB->get_records_sql("SELECT instanceid FROM {customfield_data} WHERE $fieldsql", ['fieldsearchvalue' => $fieldsearchvalue, 'fieldid' => $fieldid]);
+    }
+
+    // Sort the keys to be unique.
+    $fieldcourseids = array_pop($foundfields);
+    if (!empty($foundfields)) {
+        foreach ($foundfields as $foundfield) {
+            $fieldcourseids = array_intersect_key($fieldcourseids, $foundfield);
+            if (empty($fieldcourseids)) {
+                break;
+            }
+        }
+    }
+    if (empty($fieldcourseids)) {
+        $fieldcourseids[0] = "We didn't find any courses";
+    }
+    $courselistsql .= " AND c.id IN (" . join(',', array_keys($fieldcourseids)) . ")"; 
+}
+
 if (empty($courses)) {
     $courses = $DB->get_records_sql("SELECT ic.courseid, c.fullname FROM {iomad_courses} ic
                                                 JOIN {course} c ON (ic.courseid = c.id)
                                                 WHERE 1=1 $courselistsql
-                                                ORDER BY c.fullname");
+                                                ORDER BY c.fullname", $coursesearchparams);
 }
 
 $expirecourses = $courses;
@@ -265,17 +321,21 @@ if (!$download) {
         // Set up the filter form.
         $options = $params;
         $options['companyid'] = $companyid;
-        $mform = new iomad_user_filter_form(null, $options);
+        $mform = new \local_iomad\forms\user_search_form(null, $options);
         $mform->set_data(array('departmentid' => $departmentid));
 
         $mform->set_data($options);
         $mform->get_data();
 
         // Display the user filter form.
+        echo html_writer::start_tag('div', array('class' => 'iomadusersearchform'));
         $mform->display();
+        echo html_writer::end_tag('div');
 
         // Display the course filter form.
+        echo html_writer::start_tag('div', array('class' => 'iomadcoursesearchform'));
         $coursesform->display();
+        echo html_writer::end_tag('div');
         echo html_writer::end_tag('div');
     }
 }
