+++ local/report_completion_overview/styles.css
@@ -81,5 +81,23 @@
     align-content: end;
 }
 
+.iomadcoursesearchform .icons-collapse-expand.collapsed .expanded-icon,
+.iomadcoursesearchform .icons-collapse-expand.collapsed .collapsed-icon,
+.datesearchform  .icons-collapse-expand.collapsed .expanded-icon,
+.datesearchform .icons-collapse-expand.collapsed .collapsed-icon{
+    font-size: small;
+}
+
+.iomadcoursesearchform h3,
+.datesearchform h3 {
+    font-size: medium;
+}
+
+.iomadcoursesearchform .mform .form-inline .form-control,
+.iomadcoursesearchform .mform .form-inline .custom-select,
+.datesearchform .mform .form-inline .form-control,
+.datesearchform .mform .form-inline .custom-select {
+    max-width:75%;
+}
